flask_app/requirements.txt
# Existing dependencies (preserved from original)
requests>=2.31.0
spotipy>=2.23.0
python-dotenv>=1.0.0
clipboard>=0.0.4

# Flask dependencies for web application
Flask>=2.3.0
Flask-CORS>=4.0.0

# Additional useful dependencies
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-flask>=1.2.0
# black>=23.0.0
# flake8>=6.0.0